requirements.txt
annotated-types==0.7.0
anyio==4.11.0
blis==1.3.2
catalogue==2.0.10
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
coloredlogs==15.0.1
confection==0.1.5
cupy-cuda12x==12.3.0
cymem==2.0.13
distro==1.9.0
einops==0.8.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
fastembed==0.7.3
fastrlock==0.8.3
filelock==3.19.1
flatbuffers==25.9.23
fsspec==2025.9.0
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
langdetect==1.0.9
loguru==0.7.3
lxml==6.0.2
MarkupSafe==2.1.5
mmh3==5.2.0
mpmath==1.3.0
murmurhash==1.0.15
networkx==3.5
nltk==3.9.2
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
onnxruntime==1.23.2
openai==2.8.0
packaging==25.0
pillow==11.3.0
portalocker==3.2.0
preshed==3.0.11
protobuf==6.33.1
py_rust_stemmers==0.1.5
pydantic==2.12.4
pydantic_core==2.41.5
PyPDF2==3.0.1
python-crfsuite==0.9.11
python-docx==1.2.0
PyYAML==6.0.3
qdrant-client==1.15.1
regex==2025.11.3
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers==5.1.2
setuptools==80.9.0
six==1.17.0
slack_bolt==1.27.0
slack_sdk==3.38.0
smart_open==7.5.0
sniffio==1.3.1
spacy==3.8.9
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
sympy==1.13.1
thinc==8.3.9
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
transformers==4.57.1
triton==3.1.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
underthesea==8.3.0
underthesea_core==1.0.5
urllib3==2.5.0
wasabi==1.1.3
weasel==0.4.3
wheel==0.45.1
wrapt==2.0.1
